feat: created type file
