+++ base-components/checkbox/types.ts
@@ -1,0 +1,12 @@
+export type CheckBoxProps = {
+  title?: string;
+  description?: string;
+  checked?: boolean;
+  disabled?: boolean;
+  onCheckedChange?: (checked: boolean) => void;
+  className?: string;
+  innerClass?: string;
+  titleClass?: string;
+  descriptionClass?: string;
+  checkClass?: string;
+};
