+++ base-components/otp/types.ts
@@ -1,0 +1,6 @@
+export type OTPProps = {
+  separator?: React.ReactNode;
+  length: number;
+  value: string;
+  onChange: React.Dispatch<React.SetStateAction<string>>;
+};
