+++ base-components/smallCard/types.ts
@@ -1,0 +1,13 @@
+export interface SmallCardProps {
+  title?: string;
+  description: string;
+  icon?: React.ReactNode;
+  imageUrl?: string;
+  heading?: string;
+  className?: string;
+  buttonLabel?: string;
+  iconClassName?: string;
+  cardClassName?: string;
+  titleClassName?: string;
+  descriptionClassName?: string;
+}
