+++ base-components/smallCardWithIcon/types.ts
@@ -1,0 +1,5 @@
+export interface SmallCardWithIconType {
+  title?: string;
+  icon?: React.ReactNode;
+  iconColor?: string;
+}
