+++ base-components/timeClock/types.ts
@@ -1,0 +1,10 @@
+export interface TimeClockProps {
+  smallTitle: string;
+  largeTitle: string;
+  assigned: string;
+  groupLabel: string;
+  nameTitle: string;
+  name: string;
+  buttonText: string;
+  roles: string;
+}
