+++ components/signupPage/index.tsx
@@ -8,17 +8,18 @@ import { Input } from '@/base-components/input';
 import { Button } from '@/base-components/button';
 import SelectBox from '@/base-components/selectbox';
 import { FaGoogle, FaFacebook } from 'react-icons/fa';
+import { SignUpFormType } from './types';
 
 export default function SignUp() {
   const router = useRouter();
-  const [form, setForm] = useState({
+  const [form, setForm] = useState<SignUpFormType>({
     firstName: '',
     lastName: '',
     email: '',
     password: '',
   });
-  const [showPwd, setShowPwd] = useState(false);
-  const [touched, setTouched] = useState(false);
+  const [showPwd, setShowPwd] = useState<boolean>(false);
+  const [touched, setTouched] = useState<boolean>(false);
 
   const emailRx = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
   const pwdRx = /^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
