+++ components/signupPage/types.ts
@@ -1,0 +1,6 @@
+export type SignUpFormType = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+};
